Move CNNode in its own file
